+++ CHANGELOG.md
@@ -9,6 +9,8 @@ All notable changes to this project will be documented in this file.
 - auth group folder
 - login page route
 - login form component
+- forgot password page route
+- forgot password form component
 
 ## [0.0.0] - 2026-01-15