+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 - login form component
 - forgot password page route
 - forgot password form component
+- update password page route
+- update password form component
 
 ## [0.0.0] - 2026-01-15