+++ CHANGELOG.md
@@ -6,16 +6,52 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- auth group folder
-- login page route
-- login form component
-- forgot password page route
-- forgot password form component
-- update password page route
-- update password form component
-- dashboard group folder
+#### App
+
+- auth route group folder
+- dashboard route group folder
+
+##### Auth
+
+- login page
+- forgot password page
+- update password page
+
+##### Dashboard
+
+- dashboard page
 - dashboard layout
+
+#### Components
+
+- auth sub folder
+- dashboard sub folder
+- ui sub folder
+
+##### Auth
+
+- login form component
+- forgot password form component
+- update password form component
+
+##### Dashboard
+
 - dashboard component
+
+##### ui
+
+- shadcn components
+
+#### Hooks
+
+- hooks folder
+- shadcn use mobile
+
+#### Lib
+
+- lib folder
+- metadata constants
+- shadcn utils
 
 ## [0.0.0] - 2026-01-15