+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] - 2026-02-22
+
+### Changed
+
+- updated dependencies
+
 ## [0.1.0] - 2026-02-07
 
 ### Added